refactor(behavior): add behavior directives instead of splitting component into parent and chil components
