+++ src/demo-app/app/app.component.html
@@ -15,12 +15,12 @@
   <md-card-subtitle>Feature module</md-card-subtitle>
   <md-card-title>feature-list.component</md-card-title>
   <md-card-content>
-    <igo-feature-list-base
+    <igo-feature-list
+      igoFeatureListBehavior
       [focusFirst]="true"
-      [features]="featureService.features$ | async"
       (focus)="handleFeatureFocus($event)"
       (select)="handleFeatureSelect($event)">
-    </igo-feature-list-base>
+    </igo-feature-list>
   </md-card-content>
 
   <md-card-title>feature-details.component</md-card-title>
@@ -65,13 +65,14 @@
   <md-card-subtitle>Tool module</md-card-subtitle>
   <md-card-title>toolbar.component</md-card-title>   
   <md-card-content>
-    <igo-toolbar-base
+    <igo-toolbar
       igoToolContext
+      igoToolbarBehavior
       [withIcon]="true"
       [withTitle]="true"
       [horizontal]="true"
       (select)="handleToolSelect($event)">
-    </igo-toolbar-base>
+    </igo-toolbar>
   </md-card-content>
 
   <!--md-card-title>toolbox.component</md-card-title>
