--- src/lib/feature/feature-list/feature-list-base.component.html
@@ -1,22 +1,0 @@
-<igo-list [navigation]="true"> 
-  <ng-container *ngIf="features">
-    <ng-template
-      ngFor let-sourceFeature let-i="index"
-      [ngForOf]="features | featureGroup">
-
-      <igo-collapsible title="{{sourceFeature[0]}} ({{sourceFeature[1].length}})">
-        <ng-template ngFor let-feature let-j="index" [ngForOf]="sourceFeature[1]">
-          <igo-feature-item
-            igoListItem
-            color="accent"
-            [feature]="feature"
-            [focused]="focusFirst && i + j === 0"
-            (focus)="handleFeatureFocus(feature)"
-            (select)="handleFeatureSelect(feature)">
-          </igo-feature-item>
-        </ng-template>
-      </igo-collapsible>
-
-    </ng-template>
-  </ng-container>
-</igo-list>
