+++ src/lib/feature/feature-list/feature-list.component.spec.ts
@@ -4,10 +4,9 @@ import { IgoSharedModule } from '../../shared';
 
 import { FeatureService, FeatureGroupPipe } from '../shared';
 import { FeatureItemComponent } from '../feature-item';
-import { FeatureListBaseComponent } from './feature-list-base.component';
 import { FeatureListComponent } from './feature-list.component';
 
-describe('FeatureListComponent', () => {
+describe('FeatureListBaseComponent', () => {
   let component: FeatureListComponent;
   let fixture: ComponentFixture<FeatureListComponent>;
 
@@ -17,7 +16,6 @@ describe('FeatureListComponent', () => {
         IgoSharedModule
       ],
       declarations: [
-        FeatureListBaseComponent,
         FeatureListComponent,
         FeatureItemComponent,
         FeatureGroupPipe
