+++ src/lib/feature/feature-list/index.ts
@@ -1,2 +1,2 @@
 export * from './feature-list.component';
-export * from './feature-list-base.component';
+export * from './feature-list-behavior.directive';
