--- src/lib/filter/filterable-layer/index.ts
@@ -1,1 +1,0 @@
-export * from './filterable-layer.pipe';
