+++ src/lib/filter/module.ts
@@ -4,7 +4,7 @@ import { Md2Module } from 'md2';
 
 import { IgoSharedModule } from '../shared';
 
-import { FilterableLayerPipe } from './filterable-layer';
+import { FilterableLayerPipe } from './shared';
 import { TimeFilterFormComponent } from './time-filter-form';
 import { TimeFilterItemComponent } from './time-filter-item';
 import { TimeFilterListComponent } from './time-filter-list/';
@@ -38,7 +38,6 @@ export class IgoFilterModule {
 }
 
 export * from './shared';
-export * from './filterable-layer';
 export * from './time-filter-form';
 export * from './time-filter-item';
 export * from './time-filter-list';
