+++ src/lib/filter/shared/filterable-layer.pipe.ts
