+++ src/lib/filter/shared/index.ts
@@ -1,1 +1,2 @@
+export * from './filterable-layer.pipe';
 export * from './time-filter.interface';
