+++ src/lib/overlay/shared/index.ts
@@ -1,2 +1,3 @@
+export * from './overlay.directive';
 export * from './overlay.service';
 export * from './overlay.interface';
