+++ src/lib/query/shared/index.ts
@@ -1,1 +1,2 @@
+export * from './query.directive';
 export * from './query.service';
