+++ src/lib/query/shared/query.directive.spec.ts
