+++ src/lib/query/shared/query.directive.ts
@@ -6,7 +6,7 @@ import { MapBrowserComponent } from '../../map';
 import { Layer } from '../../layer';
 import { Feature } from '../../feature';
 
-import { QueryService } from '../shared';
+import { QueryService } from '../shared/query.service';
 
 @Directive({
   selector: '[igoQuery]'
