--- src/lib/tool/toolbar/toolbar-base.component.html
@@ -1,14 +1,0 @@
-<igo-list navigation="false"> 
-  <ng-template ngFor let-tool [ngForOf]="tools">
-    <igo-toolbar-item
-      igoListItem
-      color="accent"
-      [withTitle]="withTitle"
-      [withIcon]="withIcon"
-      [tool]=tool
-      (select)="handleToolSelect(tool)">
-    </igo-toolbar-item>
-  </ng-template>
-
-  <ng-content></ng-content>
-</igo-list>
