+++ src/lib/tool/toolbar/toolbar-behavior.directive.spec.ts
@@ -1,0 +1,14 @@
+import { TestBed } from '@angular/core/testing';
+
+describe('ToolbarBehaviorDirective', () => {
+  beforeEach(() => {
+    TestBed.configureTestingModule({
+      imports: [],
+      providers: []
+    });
+  });
+
+  it('should create an instance', () => {
+    expect(true).toBeTruthy();
+  });
+});
