+++ src/lib/tool/toolbar/toolbar.component.html
@@ -1,7 +1,14 @@
-<igo-toolbar-base
-  [tools]="tools"
-  [withTitle]="withTitle"
-  [withIcon]="withIcon"
-  [horizontal]="horizontal"
-  (select)="handleToolSelect($event)">  
-</igo-toolbar-base>
+<igo-list navigation="false"> 
+  <ng-template ngFor let-tool [ngForOf]="tools">
+    <igo-toolbar-item
+      igoListItem
+      color="accent"
+      [withTitle]="withTitle"
+      [withIcon]="withIcon"
+      [tool]=tool
+      (select)="select.emit(tool)">
+    </igo-toolbar-item>
+  </ng-template>
+
+  <ng-content></ng-content>
+</igo-list>
